chore(main): extract session-stream core from main.ts (arch R5)
Pure move of main.ts's session-stream cluster into session-stream.ts:
- createAiSdkBackendFactory(deps): BackendFactory — the whole
  backends.register('ai-sdk', …) closure plus the internal modelSupportsVision.
  Two seams resolved after the registration point are injected as accessors:
  getRuntime (SessionManager built later) and getLookupPricing (mutable pricing
  lookup, read live per recordLlmCall / snapshotted once for the lookupPricing
  field — matching the original module-let closure semantics exactly).
- createSessionStreamer(deps): StreamEvents — streamEvents + its two
  event-classifier helpers and the StreamEventsOptions/Result types.
main.ts keeps the module-scoped seams it owns (backends.register call site,
sessionActivities, lookupPricing let, runtime, the fake/e2e backend registers)
and every streamEvents call site.

Contract re-pins (maintainer-authorized): add session-stream.ts to the
main-process-contract-source-helpers aggregator; switch three direct-main.ts
readers to the combined source keeping every assertion —
attachment-frontend-contract (modelSupportsVision + supportsVision vision pin),
ipc-surface-contract (memoryPromptSnapshot + buildBackendSystemPrompt),
web-search-telemetry-scrub-contract (argsSummary scrub). registerIpc /
main-process-wiring-contract untouched.

main.ts 1656 → 1372 (−284 across both R5 clusters).
