+++ apps/desktop/src/main/__tests__/web-search-telemetry-scrub-contract.test.ts
@@ -7,18 +7,14 @@
  */
 
 import { strict as assert } from 'node:assert';
-import { readFile } from 'node:fs/promises';
 import { describe, it } from 'node:test';
-import { join, resolve } from 'node:path';
-
-const REPO_ROOT = resolve(process.cwd(), '..', '..');
+import { readMainProcessCombinedSource } from './main-process-contract-source-helpers.js';
 
 describe('WebSearch telemetry scrub contract', () => {
   it('main.ts recordToolInvocation drops argsSummary for WebSearch', async () => {
-    const src = await readFile(
-      join(REPO_ROOT, 'apps', 'desktop', 'src', 'main', 'main.ts'),
-      'utf8',
-    );
+    // The ai-sdk backend wiring (with the recordToolInvocation scrub) moved into
+    // session-stream.ts (arch R5); scan the combined main-process source.
+    const src = await readMainProcessCombinedSource();
     // Cheap grep: the wrapper that branches on toolName === WEB_SEARCH_TOOL_NAME
     // must spread the event and explicitly drop argsSummary.
     assert.match(
