+++ apps/desktop/src/main/main.ts
@@ -6,9 +6,7 @@ import { startConfigFileWatcher, type ConfigFileWatcher } from './config-file-wa
 import {
   DEFAULT_SESSION_NAME,
   filterModelVisibleTaskLedgerTasks,
-  resolveModelVisionSupport,
   DEEP_RESEARCH_SESSION_LABEL,
-  expertTeamIdFromLabels,
 } from '@maka/core';
 import type {
   BotProvider,
@@ -20,7 +18,6 @@ import type {
   SessionEvent,
 } from '@maka/core';
 import { deriveBotStatusPersistenceUpdate } from './bot-status-persistence.js';
-import { WEB_SEARCH_TOOL_NAME } from './web-search/agent-tool.js';
 import { runThreadSearch } from './search/thread-search.js';
 import { assembleDesktopTools } from './tool-assembly.js';
 import { createToolArtifactPersistence } from './tool-artifact-persistence.js';
@@ -31,22 +28,16 @@ import { CursorSubscriptionService } from './oauth/cursor-subscription-service.j
 import { AntigravitySubscriptionService } from './oauth/antigravity-subscription-service.js';
 import { importLegacyOAuthTokenFiles } from './oauth/shared-credential-bridge.js';
 import type { WorkspacePrivacyContext } from '@maka/core/incognito';
-import type { LlmCallRecord, ToolInvocationRecord } from '@maka/core/usage-stats/types';
 import { ok } from '@maka/core/settings/result';
 import {
-  AiSdkBackend,
   BackendRegistry,
   FakeBackend,
   PermissionEngine,
   SessionManager,
-  buildMcpTools,
-  buildExpertDispatchToolForTeamId,
   createLocalContinuationSafetyInspector,
   getAIModel,
   generateSessionTitle as generateRuntimeSessionTitle,
   buildProviderOptions,
-  recordLlmCall,
-  recordToolInvocation,
   buildPricingLookup,
   BotRegistry,
   setActiveProxy,
@@ -59,17 +50,11 @@ import type {
   GoalTurnOutcome,
   HostCapabilities,
   HostCapabilitiesResolver,
-  SessionActivityLease,
-  ToolArtifactRecorderInput,
-  ToolResultArchiveReaderInput,
-  ToolResultArchiveReadResult,
-  ToolResultArchiveRecorderInput,
 } from '@maka/runtime';
 import type { LlmConnection } from '@maka/core/llm-connections';
 import {
   createAgentRunStore,
   createAgentMailboxStore,
-  createAttachmentByteReader,
   createArtifactStore,
   createReadImageSnapshotter,
   createConnectionStore,
@@ -85,9 +70,7 @@ import { McpClientManager } from '@maka/mcp';
 import { registerMcpIpcMain } from './mcp-ipc-main.js';
 import {
   ensureSessionCanSendOrRebind,
-  errorCode,
   errorMessage,
-  errorReason,
   requireReadyConnection,
 } from './chat-readiness.js';
 import { createFileCredentialStore, migrateLegacyCredentials } from './credential-store.js';
@@ -103,29 +86,17 @@ import { resolveBuildInfo } from './build-info.js';
 import { OpenGatewayService } from './open-gateway.js';
 import { LocalMemoryService } from './local-memory-service.js';
 import { createAttachmentApprovalRegistry } from './attachment-approval.js';
-import {
-  buildLlmHistorySummarizer,
-  cleanupLegacyHistoryCompactArtifacts,
-  loadHistoryCompactBlocksFromArtifacts,
-  loadSynthesisCacheBlocksFromArtifacts,
-  persistSynthesisCacheBlocksToArtifacts,
-} from '@maka/runtime';
+import { cleanupLegacyHistoryCompactArtifacts } from '@maka/runtime';
 import { computerUseServiceHealth } from './computer-use-host.js';
-import {
-  computerUseAvailabilityForModel,
-  computerUseToolsForModel,
-} from './computer-use-model-tools.js';
 import { createMainWindowController } from './main-window.js';
 import { createDailyReviewMainService } from './daily-review-main.js';
 import { createPlanReminderMainService } from './plan-reminders-main.js';
 import { createBotIncomingMainService } from './bot-incoming-main.js';
 import { createSubscriptionModelFetch } from './subscription-model-fetch.js';
-import { buildDefaultContextBudgetPolicy, resolveSelectedModelContextWindow } from '@maka/runtime';
 import { createSystemPromptMainService } from './system-prompt-main.js';
 import { createMainTaskLedgerWiring } from './task-ledger-wiring.js';
 import { createMainAutomationWiring, evaluateAutomationCanFire } from './automation-wiring.js';
 import { createMainGoalWiring } from './goal-wiring.js';
-import { startDesktopSessionTurn, type SessionGoalBoundary } from './session-turn-stream.js';
 import { createOAuthModelConnectionsMainService } from './oauth-model-connections-main.js';
 import { toContractNetworkSettings } from './network-settings-main.js';
 import { registerMemoryIpc } from './memory-ipc-main.js';
@@ -151,6 +122,7 @@ import type { SettingsIpcHandle } from './settings-ipc-main.js';
 import { createVisualSmokeBotOnboardingAdapters } from './bot-onboarding-visual-smoke.js';
 import { createKeepSystemAwakeController } from './keep-system-awake.js';
 import { createSettingsRuntimeEffects } from './settings-runtime-effects.js';
+import { createAiSdkBackendFactory, createSessionStreamer } from './session-stream.js';
 import { registerGatewayIpc } from './gateway-ipc-main.js';
 import { registerSessionsIpc } from './sessions-ipc-main.js';
 import {
@@ -340,20 +312,6 @@ const planReminderStore = createPlanReminderStore(workspaceRoot);
 const taskLedgerWiring = createMainTaskLedgerWiring(workspaceRoot);
 const taskLedgerStore = taskLedgerWiring.store;
 const agentMailboxStore = createAgentMailboxStore(workspaceRoot);
-
-interface StreamEventsOptions {
-  turnId: string;
-  goalBoundary: SessionGoalBoundary;
-  activity?: SessionActivityLease;
-  observeEvent?: (event: SessionEvent) => void;
-}
-
-interface StreamEventsResult {
-  turnId: string;
-  ok: boolean;
-  error?: string;
-  outcome: GoalTurnOutcome;
-}
 
 const sessionActivities = new SessionActivityRegistry();
 
@@ -717,148 +675,33 @@ const planReminders = createPlanReminderMainService({
 
 app.setName('Maka');
 
-function modelSupportsVision(connection: LlmConnection, model: string): boolean {
-  return resolveModelVisionSupport(connection.providerType, connection.models, model);
-}
-
-backends.register('ai-sdk', async (ctx) => {
-  // MCP is optional. A corrupt mcp.json remains visible in the MCP module,
-  // but must not prevent builtin-only conversations from creating a backend.
-  await ensureMcpReady().catch(() => {});
-  const { connection, apiKey, model } = await getReadyConnection(ctx.header.llmConnectionSlug, ctx.header.model);
-  const modelFetch = buildSubscriptionModelFetch(connection, ctx.sessionId, model);
-  const memoryPromptSnapshot = await systemPromptService.buildLocalMemoryPromptFragment();
-  const supportsVision = modelSupportsVision(connection, model);
-  const candidateTools = isComputerUseRealModelE2e
-    ? computerUseTools
-    : ctx.tools
-      ? [...ctx.tools]
-      : [...builtinTools, ...buildMcpTools(mcpManager)];
-  const candidateToolAvailability = isComputerUseRealModelE2e
-    ? { economy: false, groups: [] }
-    : toolAvailability;
-  // Expert-team lead: a main session (ctx.tools undefined) labeled
-  // `mode:expert-team:<teamId>` gets the team-bound expert_dispatch tool.
-  // Child turns receive scoped `ctx.tools` and inherit the label, but must NOT
-  // get expert_dispatch — members cannot spawn nested teams.
-  const expertTeamId = ctx.tools ? undefined : expertTeamIdFromLabels(ctx.header.labels);
-  const expertDispatchTool = expertTeamId
-    ? buildExpertDispatchToolForTeamId(expertTeamId, { taskLedger: taskLedgerStore })
-    : undefined;
-  const agentTeam = ctx.agentTeam ?? (expertTeamId
-    ? { role: 'lead' as const, teamId: expertTeamId, agentId: 'lead' }
-    : undefined);
-  const backendTools = computerUseToolsForModel(
-    candidateTools,
-    computerUseTools,
-    supportsVision,
-  );
-  const backendToolAvailability = computerUseAvailabilityForModel(
-    candidateToolAvailability,
-    supportsVision,
-  );
-  const backendToolNames = new Set([
-    ...backendTools.map((tool) => tool.name),
-    ...(expertDispatchTool ? [expertDispatchTool.name, ...agentTeamLeadTools.map((tool) => tool.name)] : []),
-  ]);
-  const backendCapabilities = new Set<string>();
-  for (const [capability, tools] of [
-    ['rive', riveTools],
-    ['office', officeTools],
-    ['browser', browserTools],
-    ['computer_use', computerUseTools],
-  ] as const) {
-    if (tools.some((tool) => backendToolNames.has(tool.name))) backendCapabilities.add(capability);
-  }
-  const backendSkillHost: HostCapabilities = {
-    toolNames: backendToolNames,
-    capabilities: backendCapabilities,
-  };
-  // Child backends share the parent sessionId but intentionally have a
-  // narrower tool surface. They do not receive the Desktop Skill tool, so
-  // they must not overwrite the parent session's resolver entry.
-  if (!ctx.tools) desktopSessionSkillHosts.set(ctx.sessionId, backendSkillHost);
-
-  return new AiSdkBackend({
-    sessionId: ctx.sessionId,
-    header: { ...ctx.header, model },
-    appendMessage: ctx.appendMessage ?? ((message) => ctx.store.appendMessage(ctx.sessionId, message)),
-    connection,
-    apiKey: apiKey ?? '',
-    modelId: model,
-    permissionEngine,
-    modelFactory: (input) => getAIModel({ ...input, fetch: modelFetch }),
-    tools: expertDispatchTool
-      ? [...backendTools, expertDispatchTool, ...agentTeamLeadTools]
-      : backendTools,
-    agentTeam,
-    toolAvailability: backendToolAvailability,
-    spawnChildAgent: (input) => runtime.spawnChildAgent(ctx.sessionId, input),
-    listChildAgents: () => runtime.listChildAgents(ctx.sessionId),
-    readChildAgentOutput: (input) => runtime.readChildAgentOutput(ctx.sessionId, input),
-    providerOptions: buildProviderOptions(connection, model, ctx.header.thinkingLevel),
-    contextBudget: buildDefaultContextBudgetPolicy(connection, {
-      name: 'desktop-default-history-budget',
-      modelId: model,
-    }),
-    systemPrompt: ({ cwd }) => systemPromptService.buildBackendSystemPrompt(ctx.header, cwd, {
-      memoryFragment: memoryPromptSnapshot,
-      childInstruction: ctx.systemPrompt,
-      skillBudget: { contextWindow: resolveSelectedModelContextWindow(connection, model) },
-      host: backendSkillHost,
-    }),
-    turnTailPrompt: ({ cwd, sessionId }) => systemPromptService.buildTurnTailPrompt(cwd, sessionId),
-    shellRunContextSummary: ctx.shellRunContextSummary,
-    lookupPricing,
-    recordLlmCall: (event: LlmCallRecord) => recordLlmCall({ repo: telemetryRepo, lookupPricing }, event),
-    recordToolInvocation: (event: ToolInvocationRecord) =>
-      recordToolInvocation(
-        { repo: telemetryRepo },
-        // PR-AGENT-WEB-SEARCH-TOOL-0: scrub the query out of the
-        // telemetry record. The agent passes the raw user query as
-        // the tool argument; persisting it in `argsSummary` would
-        // leak user-derived content into the usage log.
-        event.toolName === WEB_SEARCH_TOOL_NAME
-          ? { ...event, argsSummary: undefined }
-          : event,
-      ),
-    recordToolArtifacts: (event: ToolArtifactRecorderInput) => persistToolArtifacts(ctx.header.cwd, event),
-    archiveToolResult: (event: ToolResultArchiveRecorderInput) => persistArchivedToolResult(event),
-    readToolResultArchive: (event: ToolResultArchiveReaderInput) => readArchivedToolResult(event),
-    readAttachmentBytes: createAttachmentByteReader({ artifactStore, sessionId: ctx.sessionId }),
-    ...(runtimePersistence.runtimeCommitStore
-      ? { runtimeCommitSink: runtimePersistence.runtimeCommitStore }
-      : {}),
-    supportsVision,
-    loadHistoryCompact: (event) => loadHistoryCompactBlocksFromArtifacts(artifactStore, event),
-    loadHistoryCompactCheckpoint: ctx.loadHistoryCompactCheckpoint,
-    summarizeHistoryCompact: buildLlmHistorySummarizer({
-      // Reuse the same connection/model the session already drives, so the
-      // summary stays consistent with the model that will consume it.
-      resolveModel: () =>
-        getAIModel({ connection, apiKey: apiKey ?? '', modelId: model, fetch: modelFetch }),
-      providerOptions: buildProviderOptions(connection, model, ctx.header.thinkingLevel),
-    }),
-    loadSynthesisCache: (event) => loadSynthesisCacheBlocksFromArtifacts(artifactStore, event),
-    writeSynthesisCache: (event) => persistSynthesisCacheBlocksToArtifacts(artifactStore, event, {
-      onArtifactCreated: (artifact) => {
-        safeSendToRenderer('artifacts:changed', {
-          reason: 'created',
-          artifactId: artifact.id,
-          sessionId: artifact.sessionId,
-          ts: Date.now(),
-        });
-      },
-    }),
-    recordRunTrace: ctx.recordRunTrace,
-    recordHistoryCompactCheckpoint: ctx.recordHistoryCompactCheckpoint,
-    loadTurnRuntimeEvents: ctx.loadTurnRuntimeEvents,
-    recordActiveFullCompactBlock: ctx.recordActiveFullCompactBlock,
-    recordSemanticCompactBlock: ctx.recordSemanticCompactBlock,
-    newId: randomUUID,
-    now: Date.now,
-  });
-});
+backends.register('ai-sdk', createAiSdkBackendFactory({
+  isComputerUseRealModelE2e,
+  ensureMcpReady,
+  getReadyConnection,
+  buildSubscriptionModelFetch,
+  systemPromptService,
+  mcpManager,
+  permissionEngine,
+  taskLedgerStore,
+  telemetryRepo,
+  artifactStore,
+  desktopSessionSkillHosts,
+  riveTools,
+  officeTools,
+  browserTools,
+  computerUseTools,
+  agentTeamLeadTools,
+  builtinTools,
+  toolAvailability,
+  persistToolArtifacts,
+  persistArchivedToolResult,
+  readArchivedToolResult,
+  runtimeCommitStore: runtimePersistence.runtimeCommitStore,
+  safeSendToRenderer,
+  getRuntime: () => runtime,
+  getLookupPricing: () => lookupPricing,
+}));
 
 backends.register('fake', (ctx) =>
   new FakeBackend({ sessionId: ctx.sessionId, header: ctx.header, store: ctx.store, appendMessage: ctx.appendMessage }),
@@ -1147,86 +990,15 @@ const { normalizeSettingsPatch, applySettingsRuntimeEffects, handleExternalSetti
     emitSessionsChanged,
   });
 
-function streamEvents(
-  sessionId: string,
-  iterator: AsyncIterable<SessionEvent>,
-  options: StreamEventsOptions,
-): Promise<StreamEventsResult> {
-  let userAppendBroadcasted = false;
-  const turnId = options.turnId;
-  const started = startDesktopSessionTurn({
-    sessionId,
-    events: iterator,
-    turnId,
-    goalBoundary: options.goalBoundary,
-    activities: sessionActivities,
-    ...(options.activity ? { activity: options.activity } : {}),
-    beginExternalTurn: (externalSessionId, externalTurnId) =>
-      goalWiring.coordinator.beginExternalTurn(externalSessionId, externalTurnId),
-    onEvent: (event) => {
-      if (!userAppendBroadcasted) {
-        emitSessionsChanged('message-appended', sessionId);
-        userAppendBroadcasted = true;
-      }
-      safeSendToRenderer(`sessions:event:${sessionId}`, event);
-      openGateway.publishSessionEvent(sessionId, event);
-      if (isStatusChangingSessionEvent(event)) {
-        emitSessionsChanged('status-change', sessionId);
-      }
-      if (isTurnStatusChangingSessionEvent(event)) {
-        emitSessionsChanged('turn-status-change', sessionId);
-        computerUseOverlay.clearForSession(sessionId);
-        computerUseTools.clearSession(sessionId);
-      }
-      options.observeEvent?.(event);
-    },
-    onStreamError: (error) => {
-      const event = {
-        type: 'error',
-        id: randomUUID(),
-        turnId,
-        ts: Date.now(),
-        recoverable: false,
-        code: errorCode(error),
-        reason: errorReason(error),
-        message: errorMessage(error),
-      } satisfies SessionEvent;
-      safeSendToRenderer(`sessions:event:${sessionId}`, event);
-      openGateway.publishSessionEvent(sessionId, event);
-      emitSessionsChanged('status-change', sessionId);
-      emitSessionsChanged('turn-status-change', sessionId);
-      computerUseOverlay.clearForSession(sessionId);
-      computerUseTools.clearSession(sessionId);
-    },
-    onDrained: () => {
-      emitSessionsChanged('message-appended', sessionId);
-    },
-  });
-  if (started.kind === 'unavailable') throw new Error(started.reason);
-  return started.completion.then((outcome) => {
-    const failureReason = outcome.kind === 'errored' || outcome.kind === 'suspended'
-      ? outcome.reason
-      : undefined;
-    return {
-      turnId,
-      ok: outcome.kind === 'completed',
-      ...(failureReason ? { error: failureReason } : {}),
-      outcome,
-    };
-  });
-}
-
-function isStatusChangingSessionEvent(event: SessionEvent): boolean {
-  return event.type === 'permission_request' ||
-    event.type === 'permission_decision_ack' ||
-    event.type === 'complete' ||
-    event.type === 'abort' ||
-    event.type === 'error';
-}
-
-function isTurnStatusChangingSessionEvent(event: SessionEvent): boolean {
-  return event.type === 'complete' || event.type === 'abort' || event.type === 'error';
-}
+const streamEvents = createSessionStreamer({
+  sessionActivities,
+  goalWiring,
+  openGateway,
+  computerUseOverlay,
+  computerUseTools,
+  safeSendToRenderer,
+  emitSessionsChanged,
+});
 
 async function ensureSessionCanSend(sessionId: string): Promise<void> {
   const header = await readAvailableSessionHeader(sessionId);
